+++ src/syn_net/encoding/gins.py
