+++ src/syn_net/encoding/utils.py
